+++ src/textual/box_model.py
@@ -124,7 +124,7 @@ def get_box_model(
         )
         content_height = min(content_height, max_height)
 
-    content_height = max(Fraction(1), content_height)
+    content_height = max(Fraction(0), content_height)
     model = BoxModel(
         content_width + gutter.width, content_height + gutter.height, margin
     )